test(e2e): harden db-free runner
- ✅ Run the production Cloudflare build through Wrangler without database credentials
- 🧪 Block unmocked API requests and split the Gear Library scenarios
- 🔧 Use Vite+ locally while keeping GitHub Actions on pnpm
- 📦 Add vue-router 5.1.0 as a direct dependency
