+++ tests/playwright/dashboard/sidebar-collapsed.test.ts
@@ -14,6 +14,12 @@ async function mockGuestLogin(context: BrowserContext) {
 }
 
 async function mockGearLibraryReads(context: BrowserContext) {
+  await context.route('**/api/equipment/brands**', async (route) => {
+    await route.fulfill({
+      json: []
+    })
+  })
+
   await context.route('**/api/equipment/categories**', async (route) => {
     await route.fulfill({
       json: []
