+++ tests/playwright/runtime.env
@@ -1,0 +1,6 @@
+NUXT_DATABASE_URL=
+NUXT_LOCAL_DATABASE=
+NUXT_OAUTH_TWITCH_CLIENT_ID=e2e-client
+NUXT_OAUTH_TWITCH_CLIENT_SECRET=e2e-secret
+NUXT_SESSION_SECRET=e2e-session-secret-with-at-least-32-characters
+PERD_E2E_PAGES=true
